+++ next.config.ts
@@ -1,7 +1,20 @@
 import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
-  /* config options here */
+  images: {
+    remotePatterns: [
+      {
+        protocol: "https",
+        hostname: "assets-bucket.deadlock-api.com",
+        pathname: "/**",
+      },
+      {
+        protocol: "https",
+        hostname: "assets.deadlock-api.com",
+        pathname: "/**",
+      },
+    ],
+  },
 };
 
 export default nextConfig;